change default behavior from full build to quick build including command line flags
Up to now, the default of ibutex was doing a full build with multiple runs of the LaTeX-Engine and the bibtex engine.
However, most of the time during development of either a thesis or a presentation, a full run is usually not necessary.
Therefore ibutex now by default runs only once, dropping the -q/--quick flag in the process, introducing the -f/--full
flag instead to issue a full run.